--- Sources/Old/CoordinatorRouter.swift
@@ -1,16 +1,0 @@
-//
-//  CoordinatorRouter.swift
-//  CoordinatorKit-iOS
-//
-//  Created by Bruno Fernandes on 9/15/18.
-//  Copyright © 2018 bfernandesbfs. All rights reserved.
-//
-
-import UIKit
-
-public protocol CoordinatorRouter: Presentable {
-
-    var rootViewController: UIViewController? { get }
-
-    var router: RouterType? { get set }
-}
