--- Sources/Old/Router/Presentable.swift
@@ -1,19 +1,0 @@
-//
-//  Presentable.swift
-//  CoordinatorKit-iOS
-//
-//  Created by Bruno Fernandes on 9/15/18.
-//  Copyright © 2018 bfernandesbfs. All rights reserved.
-//
-
-import UIKit
-
-public protocol Presentable {
-    func toPresentable() -> UIViewController
-}
-
-extension UIViewController: Presentable {
-    public func toPresentable() -> UIViewController {
-        return self
-    }
-}
