--- Tests/Old/Coordinator/CoordinatorRouterSpec.swift
@@ -1,27 +1,0 @@
-//
-//  CoordinatorRouterSpec.swift
-//  CoordinatorKit
-//
-//  Created by Bruno Fernandes on 9/16/18.
-//  Copyright © 2018 bfernandesbfs. All rights reserved.
-//
-
-import Quick
-import Nimble
-@testable import CoordinatorKit
-
-class CoordinatorRouterSpec: QuickSpec {
-
-    override func spec() {
-
-        describe("CoordinatorRouterSpec") {
-
-            beforeEach {
-            }
-
-            it("work") {
-            }
-        }
-    }
-}
-
